Добавляет функцию для генерации фото, выносит данные в отдельный модуль, правки
