+++ js/comments.js
@@ -1,41 +1,5 @@
 import {getRandomInteger} from './util.js';
-
-const massageList = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-
-const nameList = [
-  'Илларион',
-  'Аглая',
-  'Елисей',
-  'Василиса',
-  'Платон',
-  'Серафима',
-  'Тихон',
-  'Лукерья',
-  'Демид',
-  'Фекла',
-  'Арсений',
-  'Алевтина',
-  'Захар',
-  'Доминика',
-  'Прохор',
-  'Евдокия',
-  'Ефим',
-  'Ангелина',
-  'Макар',
-  'Пелагея',
-  'Родион',
-  'Злата',
-  'Дмитрий',
-  'Мария',
-  'Павел'
-];
+import {massageList, nameList} from './data.js';
 
 const MIN_ID_COMMENT = 1;
 const MAX_ID_COMMENT = 1000;
