+++ js/data.js
@@ -1,7 +1,4 @@
-import {getRandomInteger} from './util.js';
-import {createComment} from './comments.js';
-
-const descriptionList = [
+const descriptionPhotoList = [
   'Просто наслаждаюсь моментом.',
   'Солнечные дни и счастливые улыбки.',
   'Ловлю лучи солнца на лице.',
@@ -29,36 +26,41 @@ const descriptionList = [
   'Просто люблю жизнь.'
 ];
 
-const MIN_DESCRIPTION_INDEX = 1;
+const massageList = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
 
-const MIN_LIKES = 15;
-const MAX_LIKES = 200;
+const nameList = [
+  'Илларион',
+  'Аглая',
+  'Елисей',
+  'Василиса',
+  'Платон',
+  'Серафима',
+  'Тихон',
+  'Лукерья',
+  'Демид',
+  'Фекла',
+  'Арсений',
+  'Алевтина',
+  'Захар',
+  'Доминика',
+  'Прохор',
+  'Евдокия',
+  'Ефим',
+  'Ангелина',
+  'Макар',
+  'Пелагея',
+  'Родион',
+  'Злата',
+  'Дмитрий',
+  'Мария',
+  'Павел'
+];
 
-let id = 1;
-let urlFotoIndex = 1;
-
-const MIN_COUNT_COMMENTS = 0;
-const MAX_COUNT_COMMENTS = 30;
-
-const COUNT_FOTO = 25;
-
-//функция генерации объектов-описаний фотографий
-const createDescriptionFoto = () => {
-  const randomDescriptionIndex = getRandomInteger(MIN_DESCRIPTION_INDEX, descriptionList.length - 1);
-  const randomLikesIndex = getRandomInteger(MIN_LIKES, MAX_LIKES);
-  const randomCountComments = getRandomInteger(MIN_COUNT_COMMENTS, MAX_COUNT_COMMENTS);
-  const сomments = Array.from({length: randomCountComments}, createComment);
-
-  return {
-    id: id++,
-    url: `photos/${ urlFotoIndex++ }.jpg`,
-    description: descriptionList[randomDescriptionIndex],
-    likes: randomLikesIndex,
-    comments: сomments
-  };
-};
-
-const generateDescriptionFoto = () => Array.from({length: COUNT_FOTO}, createDescriptionFoto);
-
-export {generateDescriptionFoto};
-
+export {descriptionPhotoList, massageList, nameList};
